+++ src/main.rs
@@ -22,8 +22,8 @@ fn main() {
 
     let meterial_ground: Arc<dyn Material> = Arc::new(material::Lambertian::new(Color::new(0.5, 0.5, 0.8)));
     let material_center: Arc<dyn Material> = Arc::new(material::Lambertian::new(Color::new(0.1, 0.2, 0.5)));
-    let material_left: Arc<dyn Material> = Arc::new(material::Dielectric::new(1.0 / 1.33));
-    // let material_bubble: Arc<dyn Material> = Arc::new(material::Dielectric::new(1.0 / 1.50));
+    let material_left: Arc<dyn Material> = Arc::new(material::Dielectric::new(1.5));
+    let material_bubble: Arc<dyn Material> = Arc::new(material::Dielectric::new(1.0 / 1.50));
     let material_right: Arc<dyn Material> = Arc::new(material::Metal::new(Color::new(0.8, 0.6, 0.2), 1.0));
 
     world.add(Box::new(
@@ -35,9 +35,9 @@ fn main() {
     world.add(Box::new(
         Sphere::new(Point3::new(-1.0, 0.0, -1.0), 0.5, Arc::clone(&material_left))
     ));
-    // world.add(Box::new(
-    //     Sphere::new(Point3::new(-1.0, 0.0, -1.0), 0.4, Arc::clone(&material_bubble))
-    // ));
+    world.add(Box::new(
+        Sphere::new(Point3::new(-1.0, 0.0, -1.0), 0.4, Arc::clone(&material_bubble))
+    ));
     world.add(Box::new(
         Sphere::new(Point3::new(1.0, 0.0, -1.0), 0.5, Arc::clone(&material_right))
     ));